Add gorm plugin for encrypted fields
This adds a gorm plugin that will automatically encrypt/decrypt
fields if they have `silo:"encrypted"` tags, using the `silo:"aad"`
tags as additional data.  If encryption/decryption fails, the plugin
will panic so that further processing for writing/returning data
does not occur to prevent accidental writing of encrypted data.

The plugin is added to the db object and initialized within
keystore so it is fully transparent to the rest of the application.